Just discovered (!) iCloud and Core Data - Syncing Data - objc.io issue #10 http://www.objc.io/issue-10/icloud-core-data.html
I modified the observers to follow (neglecting for this commit  -migrateiCloudStoreToLocalStore and “To set up your persistent store for iCloud, there are a few options you need to be aware of to pass when calling addPersistentStoreWithType:configuration:URL:options:error: or migratePersistentStore:toURL:options:withType:error::”

This version works but badly. It often crashes on a Core Data error. But I was able to demonstrate a sustained ping exchange.

Next: -addPersistentStoreWithType: to move the document to iCloud.